generate float data instead of int

+++ data.py
@@ -21,7 +21,7 @@ spacing = float(sys.argv[4])
 #         )
 
 f = h5py.File("data.hdf5")
-d = f.create_dataset('transform', (count*chunks, ndim), int, chunks=(chunks, 1))
+d = f.create_dataset('transform', (count*chunks, ndim), float, chunks=(chunks, 1))
 d[...] = numpy.random.random((count*chunks, ndim))*spacing
 #d[...] = coords
 f.close()
